fix: make function-passing to fa-transform fields more reasonable
Previously, when a function was passed to a fa-modifier transform field
(like `fa-rotate`), `fa-modifier` expected that function to return a
full Transform array, while the most useful and expected case would be
to return the values approriate to that specific field (e.g.
for `fa-rotate="myFunc"`, `myFunc()` should reasonably return
an array [x, y, z].  Now this is the supported case.